+++ src/javascript/EditPanel/header/HeaderLowerSection.scss
