+++ src/javascript/EditPanel/header/HeaderLowerSection.spec.js
