+++ src/javascript/EditPanel/header/HeaderUpperSection.scss
