+++ src/javascript/EditPanel/header/HeaderUpperSection.spec.js
